Add a public variant format
Will be used to identify public versions of documents in the Archive.
A public version of a document is an additional variant of a document
in addition to the production and archival versions/variants.
Typically, the public version will be based on the latest archival
version of a document. A public version of a document will typically
contain redacted text and there might be pages omitted.